requirements.txt
# Global ----------------------------------------------------------------------
LRFutils   == 0.1.2
numpy
matplotlib
tensorflow==2.10.0
#tensorrt
scipy
pyvtk
multiprocess
astropy
bokeh
pyyaml
pyjson
pandas
GPUtil
psutil
gitpython
